Add workaround for smooth scrolling in Vim
I've tried using every possible autocommand buffer event to call SexyScroller_ScrollToCursor(0)
correctly, without avail. Each time you the call to winsaveview() in that function returns the
incorrect window view information, e.g., you would expect calling winsaveview in the BufEnter
event to give the new view when switching buffers, but when you do :bn or :bp, the incorrect
view is returned for the buffer that is loaded into the window. So I gave up on that route and
in this workaround I simply call the function before each scrolling key-combination, which
unfortunately results in double function calls.